requirements-sql.txt
pytest
flask
flask_security
flask_sqlalchemy
bcrypt
uwsgi
pylint
email_validator
werkzeug
